New provision.sh called from vagrant so integration5.0snapshot.sh doesn't have hard-coded QADIR. Many other improvements so that provision can be run multiple times.

+++ qa/check.sh
@@ -18,16 +18,34 @@ dpkg --list $PRODUCTS
 echo -e "\n `date`-----------------Service Status---------------------------------------"
 for i in $PRODUCTS; do service $i status; done
 
-rm index.html*
-# try 10 times, 2 seconds apart
+if [ -f index.html]; then rm index.html*; fi
+# try 20 times, 4 seconds apart
 echo -e "\n `date`------------wait for Elasticserach to be up----------------------------------------"
 for i in `seq 1 20`; do echo "wget es index" && sleep 4 && wget -q --http-user=$ELASTICUSER --http-password=$ELASTICPWD http://localhost:9200 && break; done
 cat index.html
 
-echo -e "\n `date`----------- Wait for elasticsearch and kibana plugin status to be green ----------------"
+echo -e "\n `date`----------- Wait for elasticsearch plugin status to be green ----------------"
 for i in `seq 1 30`; do echo "${i} `date` grep kibana log for 'green'" && sleep 4 && grep '"plugin:elasticsearch@.*","info"],"pid":.*,"state":"green"' /var/log/kibana/kibana.stdout && break; done
+if ! grep '"plugin:elasticsearch@.*","info"],"pid":.*,"state":"green"' /var/log/kibana/kibana.stdout
+  then (
+    echo "\n `date`----------- FAILED to find plugin:elasticsearch@.* state:green in /var/log/kibana/kibana.stdout"
+    echo "-------- Found:"
+    cat /var/log/kibana/kibana.stdout
+    exit 1
+  )
+fi
+
+echo -e "\n `date`----------- Wait for kibana plugin status to be green ----------------"
 for i in `seq 1 30`; do echo "${i} `date` grep kibana log for 'green'" && sleep 4 && grep '"plugin:kibana@.*","info"],"pid":.*,"state":"green"' /var/log/kibana/kibana.stdout && break; done
 #for i in `seq 1 20`; do echo "grep kibana log for 'green'" && sleep 4 &&  journalctl -u kibana.service --since "5 minutes ago"| grep "plugin:elasticsearch.*green" && break; done
+if ! grep '"plugin:elasticsearch@.*","info"],"pid":.*,"state":"green"' /var/log/kibana/kibana.stdout
+  then (
+    echo "\n `date`----------- FAILED to find plugin:kibana@.* state:green in /var/log/kibana/kibana.stdout"
+    echo "-------- Found:"
+    cat /var/log/kibana/kibana.stdout
+    exit 1
+  )
+fi
 
 
 echo -e "\n `date`----------- Status of each kibana plugin ------------------------"
@@ -38,8 +56,6 @@ done
 
 #journalctl --since "2016-03-30"
 #journalctl -u kibana.service
-#echo -e "\n----------------------------------------------------------------------"
-#/usr/share/elasticsearch/bin/xpack/esusers useradd admin -r admin -p notsecure
 
 echo -e "\n `date`-----------------Shield File Users-----------------------------------------"
 /usr/share/elasticsearch/bin/x-pack/users list
